requirements.txt
streamlit
python-dotenv
pdf2image
google-generativeai
PyPDF2
pandas
pdfplumber
spacy
scikit-learn
nltk
